(PC-24812) fix(sonar): fix code smell
